+++ docs/docs_server.sh
@@ -1,3 +1,4 @@
 #!/usr/bin/env bash
 
-sphinx-autobuild docs build/html
+sphinx-apidoc -f -o docs/source/ freckles
+sphinx-autobuild -H 0.0.0.0 -p 8082 docs build/html