+++ freckles/freckles_defaults.py
@@ -51,13 +51,17 @@ ADAPTER_MARKER_EXTENSION = "adapter.freckle"
 ADAPTER_TASKS_EXTENSION = "tasks.adapter.freckle"
 ADAPTER_INIT_EXTENSION = "init.adapter.freckle"
 
-BLUEPRINT_URL_PREFIX = "blueprint:"
+BLUEPRINT_URL_PREFIX = "blueprint"
+BLUEPRINT_DEFAULTS_URL_PREFIX = "blueprint-defaults"
 BLUEPRINT_MARKER_EXTENSION = "blueprint.freckle"
 
 DEFAULT_FRECKLE_TARGET_MARKER = "__auto__"
 
 DEFAULT_FRECKELIZE_PROFILE_PRIORITY = 100
 FRECKELIZE_PROFILE_ACTIVE_KEY = "freckle_profile_active"
+
+DEFAULT_FRECKELIZE_TARGET_FOLDER = "~/freckles"
+
 
 DEFAULT_REPOS = {
     "default": {"roles": [(None, defaults.DEFAULT_ROLES_PATH)],